Refactor airlift defintions construction logic for more elegant replacement
