+++ examples/experimental/dagster-airlift/dagster_airlift/core/load_defs.py
@@ -7,6 +7,8 @@ from dagster import (
     _check as check,
     external_asset_from_spec,
 )
+from dagster._core.definitions.definitions_loader import DefinitionsLoadContext, DefinitionsLoadType
+from dagster._serdes.serdes import deserialize_value, serialize_value
 from dagster._utils.warnings import suppress_dagster_warnings
 
 from dagster_airlift.constants import DAG_ID_METADATA_KEY, TASK_ID_METADATA_KEY
@@ -16,10 +18,28 @@ from dagster_airlift.core.sensor import (
     DEFAULT_AIRFLOW_SENSOR_INTERVAL_SECONDS,
     build_airflow_polling_sensor,
 )
-from dagster_airlift.core.serialization.compute import (
-    defs_with_airflow_metadata_applied,
-    get_or_compute_airflow_data,
-)
+from dagster_airlift.core.serialization.compute import _metadata_key, compute_serialized_data
+
+
+def get_or_compute_airflow_data(
+    *, airflow_instance: AirflowInstance, defs: Definitions
+) -> AirflowDefinitionsData:
+    context = DefinitionsLoadContext.get()
+    metadata_key = _metadata_key(airflow_instance.name)
+    if (
+        context.load_type == DefinitionsLoadType.RECONSTRUCTION
+        and metadata_key in context.reconstruction_metadata
+    ):
+        airflow_definition_data = deserialize_value(
+            context.reconstruction_metadata[metadata_key], as_type=AirflowDefinitionsData
+        )
+        assert isinstance(airflow_definition_data, AirflowDefinitionsData)
+        return airflow_definition_data
+    else:
+        serialized_data = compute_serialized_data(airflow_instance=airflow_instance, defs=defs)
+        return AirflowDefinitionsData(
+            instance_name=airflow_instance.name, serialized_data=serialized_data
+        )
 
 
 @suppress_dagster_warnings
@@ -33,6 +53,8 @@ def build_defs_from_airflow_instance(
     airflow_data = get_or_compute_airflow_data(airflow_instance=airflow_instance, defs=defs)
     return definitions_from_airflow_data(
         airflow_data, defs, airflow_instance, sensor_minimum_interval_seconds
+    ).with_reconstruction_metadata(
+        {_metadata_key(airflow_data.instance_name): serialize_value(airflow_data)}
     )
 
 
@@ -67,18 +89,15 @@ def defs_with_assets_and_sensor(
         minimum_interval_seconds=sensor_minimum_interval_seconds,
         airflow_data=airflow_data,
     )
-    return defs_with_airflow_metadata_applied(
-        Definitions(
-            assets=assets_defs,
-            asset_checks=defs.asset_checks if defs else None,
-            sensors=[airflow_sensor, *defs.sensors] if defs and defs.sensors else [airflow_sensor],
-            schedules=defs.schedules if defs else None,
-            jobs=defs.jobs if defs else None,
-            executor=defs.executor if defs else None,
-            loggers=defs.loggers if defs else None,
-            resources=defs.resources if defs else None,
-        ),
-        airflow_data,
+    return Definitions(
+        assets=assets_defs,
+        asset_checks=defs.asset_checks if defs else None,
+        sensors=[airflow_sensor, *defs.sensors] if defs and defs.sensors else [airflow_sensor],
+        schedules=defs.schedules if defs else None,
+        jobs=defs.jobs if defs else None,
+        executor=defs.executor if defs else None,
+        loggers=defs.loggers if defs else None,
+        resources=defs.resources if defs else None,
     )
