+++ examples/experimental/dagster-airlift/dagster_airlift_tests/unit_tests/core_tests/test_load_defs.py
@@ -22,6 +22,7 @@ from dagster._core.test_utils import environ
 from dagster_airlift.core import (
     build_defs_from_airflow_instance as build_defs_from_airflow_instance,
 )
+from dagster_airlift.core.airflow_defs_data import AirflowDefinitionsData
 from dagster_airlift.core.serialization.compute import compute_serialized_data
 from dagster_airlift.test import make_instance
 from dagster_airlift.utils import DAGSTER_AIRLIFT_MIGRATION_STATE_DIR_ENV_VAR
@@ -323,6 +324,9 @@ def test_cached_loading() -> None:
     } == {a, AssetKey(["airflow_instance", "dag", "dag"])}
     assert len(defs.metadata) == 1
     assert "dagster-airlift/source/test_instance" in defs.metadata
+    assert isinstance(
+        defs.metadata["dagster-airlift/source/test_instance"].value, AirflowDefinitionsData
+    )
 
     # Create a load definitions_load_context with cache data
     context_with_cache = DefinitionsLoadContext(
